% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_advice_banner.R
\name{hr_advice_banner}
\alias{hr_advice_banner}
\title{Advice sheet banner}
\usage{
hr_advice_banner(
  year_end,
  tac,
  tac_last_year,
  note = "",
  publication_date,
  period = "fishing_year",
  publication_note = ""
)
}
\arguments{
\item{year_end}{the assessment year}

\item{tac}{TAC}

\item{tac_last_year}{last years tac}

\item{note}{banner note, if needed}

\item{publication_date}{\code{BigD::fdt()}-parsable date string with the publication date}

\item{period}{Advisory period. Default is 'fishing_year', other options: 'annual', '2year', '3year', '5year'}

\item{publication_note}{publication note, if needed}
}
\description{
Advice sheet banner
}
