% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hr_astand_theme}
\alias{hr_astand_theme}
\title{The Ástand theme}
\usage{
hr_astand_theme(...)
}
\arguments{
\item{...}{input to ggplot theme}
}
\value{
ggplot2 theme for the Ástand
}
\description{
Ástand functions
}
