% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hr_astand_x_scale}
\alias{hr_astand_x_scale}
\title{Astand x scale}
\usage{
hr_astand_x_scale(x, tyr = lubridate::year(Sys.Date()), ...)
}
\arguments{
\item{x}{break point}

\item{tyr}{terminal year}

\item{...}{Further input to ggplot2::scale_x_continuous()}
}
\description{
Astand x scale
}
