% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hr_red_dot_number}
\alias{hr_red_dot_number}
\title{Print number with a thousand separator}
\usage{
hr_red_dot_number(number)
}
\arguments{
\item{number}{the number you want to convert}
}
\description{
Print number with a thousand separator
}
