% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_techreport_plot_landings_country.R
\name{hr_techreport_plot_landings_country}
\alias{hr_techreport_plot_landings_country}
\title{Plot landings by country}
\usage{
hr_techreport_plot_landings_country(
  pcon,
  ylab = hr_label("landings_kt"),
  xlab = hr_label("year"),
  breaks = seq(0, 1e+05, by = 10),
  year_start = 1000,
  year_end = 9999
)
}
\arguments{
\item{pcon}{A database connection object compatible with \code{dplyr::tbl}.}

\item{ylab}{Character. Y-axis label. Defaults to the localised label for
landings in thousands of tonnes.}

\item{xlab}{Character. X-axis label. Defaults to the localised label for
year.}

\item{breaks}{Numeric vector of x-axis break points. Default is every
10 units from 0 to 100 000.}

\item{year_start}{Integer. First year to include. Default is \code{1000}.}

\item{year_end}{Integer. Last year to include. Default is \code{9999}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Queries the \code{landings} table and creates a stacked bar chart of
annual landings in thousands of tonnes, split between Icelandic and
foreign catches. Country labels are localised to the current language.
}
