% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_techreport_plot_numboats.R
\name{hr_techreport_plot_numboats}
\alias{hr_techreport_plot_numboats}
\title{Plot number of vessels accounting for 95\% of catch}
\usage{
hr_techreport_plot_numboats(
  pcon,
  year_start = 1000,
  year_end = 9999,
  landings_tbl = dplyr::tbl(pcon, "landings"),
  ices_area_like = NULL,
  country = NULL
)
}
\arguments{
\item{pcon}{A database connection object compatible with \code{dplyr::tbl}.}

\item{year_start}{Integer. First year to include. Default is \code{1000}.}

\item{year_end}{Integer. Last year to include. Default is \code{9999}.}

\item{landings_tbl}{The landings to use, e.g. without released fish.
Default the \code{landings} table of \code{pcon}.}

\item{ices_area_like}{SQL LIKE pattern of the ICES areas of the landings,
e.g. \code{"5a\%"}. Default \code{NULL}, all areas.}

\item{country}{Countries of the landings, e.g. \code{"Iceland"}. Default
\code{NULL}, all.}
}
\value{
A \code{patchwork} plot object.
}
\description{
Queries the \code{landings} table and computes the minimum number of
vessels responsible for 95\\% of total catch in each year. Returns a
two-panel \code{patchwork} plot: a time series (left) and a
catch-vs-number-of-vessels phase plot (right).
}
