% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_theme_crayola.R
\name{hr_theme_crayola_col}
\alias{hr_theme_crayola_col}
\title{crayola colors}
\usage{
hr_theme_crayola_col(n = 100, ...)
}
\arguments{
\item{n}{Defines Length of color entries in the vector of palettes is calculated as n*12 (this length should be longer than levels filled).}

\item{...}{Additional arguments passed to ggplot2::scale_fill_manual.}
}
\value{
Functions the same as ggplot2::scale_fill_manual but with color palette vector defined according to the 'crayola' spectrum
}
\description{
crayola colors
}
