% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_panel_plot.R
\name{two_panel_plot}
\alias{two_panel_plot}
\title{Two panel plot}
\usage{
two_panel_plot(
  dat,
  prop.text = hr_label("prop_of_measure"),
  total.text = "\%s (kt)",
  y = hr_label("catches"),
  x = hr_label("year"),
  fill = "",
  cols = c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00",
    "#CC79A7", "black"),
  split_by_gear = FALSE
)
}
\arguments{
\item{dat}{input data frame}

\item{prop.text}{Proportion text}

\item{total.text}{total text}

\item{y}{type of data label}

\item{x}{Time label}

\item{fill}{Fill label}

\item{cols}{fill colours}

\item{split_by_gear}{Boolean, Use mfdb_gear_code as a facet?}
}
\value{
ggplot object
}
\description{
Two panel plot
}
