% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_techreport_plot_landings_gear.R
\name{hr_techreport_plot_landings_gear}
\alias{hr_techreport_plot_landings_gear}
\title{Plot Icelandic landings by gear (total and stacked)}
\usage{
hr_techreport_plot_landings_gear(pcon, year_start = 1000, year_end = 9999)
}
\arguments{
\item{pcon}{A database connection object compatible with \code{dplyr::tbl}.}

\item{year_start}{Integer. First year to include. Default is \code{1000}.}

\item{year_end}{Integer. Last year to include. Default is \code{9999}.}
}
\value{
A \code{ggplot2} / \code{patchwork} plot object.
}
\description{
Queries the \code{landings} table of a pax database, groups by gear
(BMT, DSE, LLN, Other), and creates a two-panel plot showing total
Icelandic landings (in thousands of tonnes) and their proportional
composition by gear over time.
}
