% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_techreport_plot_numboats.R
\name{hr_techreport_plot_numboats}
\alias{hr_techreport_plot_numboats}
\title{Plot number of vessels accounting for 95\% of catch}
\usage{
hr_techreport_plot_numboats(pcon, year_start = 1000, year_end = 9999)
}
\arguments{
\item{pcon}{A database connection object compatible with \code{dplyr::tbl}.}

\item{year_start}{Integer. First year to include. Default is \code{1000}.}

\item{year_end}{Integer. Last year to include. Default is \code{9999}.}
}
\value{
A \code{patchwork} plot object.
}
\description{
Queries the \code{landings} table and computes the minimum number of
vessels responsible for 95\\% of total catch in each year. Returns a
two-panel \code{patchwork} plot: a time series (left) and a
catch-vs-number-of-vessels phase plot (right).
}
