% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_techreport_table_boatsummary.R
\name{hr_techreport_table_boatsummary}
\alias{hr_techreport_table_boatsummary}
\title{Format vessel and catch summary table by gear}
\usage{
hr_techreport_table_boatsummary(pcon, year_start = 1000, year_end = 9999)
}
\arguments{
\item{pcon}{A database connection object compatible with \code{dplyr::tbl}.}

\item{year_start}{Integer. First year to include. Default is \code{1000}.}

\item{year_end}{Integer. Last year to include. Default is \code{9999}.}
}
\value{
A \code{gt} table object.
}
\description{
Queries the \code{landings} table, groups by gear, and produces a GT table
summarising the number of vessels and catch (in thousands of tonnes) per
gear type per year. Column headers are localised to the current language.
}
